refactor: move common bus types into bus::types module
