+++ src/swapd/swap_state.rs
@@ -1,8 +1,9 @@
 use farcaster_core::role::{SwapRole, TradeRole};
 use strict_encoding::{StrictDecode, StrictEncode};
 
-use crate::bus::ctl::{Outcome, Params};
+use crate::bus::ctl::Params;
 use crate::bus::msg::Commit;
+use crate::bus::Outcome;
 
 #[derive(Display, Debug, Clone, StrictEncode, StrictDecode)]
 pub enum AliceState {
